+++ tests/test_schunk_insert.py
@@ -42,7 +42,7 @@ def test_schunk_insert_numpy(contiguous, urlpath, nchunks, ninserts, copy, creat
 
     for i in range(ninserts):
         pos = random.randint(0, nchunks + i)
-        buffer = pos * numpy.arange(200 * 1000)
+        buffer = pos * numpy.arange(200 * 1000, dtype="int32")
         if create_chunk:
             chunk = blosc2.compress2(buffer)
             schunk.insert_chunk(pos, chunk)
